Code cleanup, docs
