+++ FlintCore/Actions/TerminatingAction.swift
@@ -11,6 +11,9 @@ import Foundation
 /// An action that performs no work except successfully completing with feature termination.
 ///
 /// Actions adopting this protocol will not need to provide a `perform` implementation.
+///
+/// Conforming to this protocol is useful for "done" type Actions that you want to participate
+/// in standard Action patterns, but do not actually perform any code.
 public protocol TerminatingAction: Action {
 }
 
@@ -19,25 +22,3 @@ extension TerminatingAction {
         return completion.completedSync(.successWithFeatureTermination)
     }
 }
-
-#if os(iOS)
-public struct DismissInput: FlintLoggable {
-    public let animated: Bool
-    
-    public static func animated(_ animated: Bool) -> DismissInput {
-        return DismissInput(animated: animated)
-    }
-}
-
-public protocol DismissingUIAction: UIAction {
-    typealias InputType = DismissInput
-    typealias PresenterType = UIViewController
-}
-
-extension DismissingUIAction {
-    public static func perform(context: ActionContext<InputType>, presenter: PresenterType, completion: Completion) -> Completion.Status {
-        presenter.dismiss(animated: context.input.animated)
-        return completion.completedSync(.successWithFeatureTermination)
-    }
-}
-#endif
